refactor: 특정 컴포넌트에서만 쓰이는 상태및 함수는 해당 컴포넌트로 이동
